Add goToTarget() method for Creature
The method moves the creature along the path to the target
by a number of cells that depends on the speed. If an obstacle appears
on the way or the target disappears, it looks for a new path and moves
along it. If there is no target, it stays in place.