Refactor 'path' to final and findPath() to void
This will allow us to get rid of null checks for the 'path' field. We
will store a single instance of Deque, which we will pass to
PathFinder.findP() as a parameter, and we will add the path to it.
Also, get rid of passing WorldMap as a parameter to pathFind(), since we
 have only one instance of WorldMap in the entire program, and we will
 take it from WorldMap.getInstance().